Support categories in OPML import

+++ app/Actions/ImportOPML.php
@@ -7,6 +7,7 @@ namespace App\Actions;
 use App\Actions\Feed\CreateNewFeed;
 use App\Models\EntryInteraction;
 use App\Models\FeedSubscription;
+use App\Models\SubscriptionCategory;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
@@ -26,19 +27,26 @@ class ImportOPML
     {
         $file = $request->file('opml_file');
 
-        $xml = simplexml_load_file($file);
+        if (! $file || ! ($xml = simplexml_load_file($file->getPathname()))) {
+            throw new \Exception('Unable to parse OPML file');
+        }
 
         // TODO: make this optional
         EntryInteraction::where('user_id', Auth::user()->id)->delete();
         FeedSubscription::where('user_id', Auth::user()->id)->delete();
 
-        foreach ($xml->body->outline as $category) {
-            foreach ($category->outline as $outline) {
-                $feed_url = (string) $outline['xmlUrl'];
+        foreach ($xml->body->outline as $category_outline) {
+            foreach ($category_outline->outline as $feed_outline) {
+                $feed_url = (string) $feed_outline['xmlUrl'];
+
+                $category = SubscriptionCategory::firstOrCreate([
+                    'user_id' => Auth::user()->id,
+                    'name' => (string) $category_outline['text'],
+                ]);
 
                 Log::info("[OPML] Importing feed: {$feed_url} for user: ".Auth::user()->id);
 
-                CreateNewFeed::dispatch($feed_url, Auth::user());
+                CreateNewFeed::dispatch($feed_url, Auth::user(), $category->id);
             }
         }
